Move the AddBundleCompletionProposal and BuildpathQuickFixProcessor
Now we have more and more features in this bundle it seems good to
structure its packages more as we have already done for the other
imported parts.

This moves the bundle completion proposal and buildpath quickfix to
their own packages and removes the export of org.eclipse.pde.bnd.ui what
should only be used internally.
