+++ ui/org.eclipse.pde.bnd.ui/META-INF/MANIFEST.MF
@@ -4,8 +4,9 @@ Bundle-Name: Generic UI components related to BND
 Bundle-SymbolicName: org.eclipse.pde.bnd.ui;singleton:=true
 Bundle-Vendor: Eclipse.org
 Bundle-Version: 1.0.0.qualifier
-Export-Package: org.eclipse.pde.bnd.ui;version="1.0.0";x-friends:="org.eclipse.pde.ui",
+Export-Package: org.eclipse.pde.bnd.ui.autocomplete;version="1.0.0";x-friends:="org.eclipse.pde.ui",
  org.eclipse.pde.bnd.ui.preferences;version="1.0.0";x-friends:="org.eclipse.pde.ui",
+ org.eclipse.pde.bnd.ui.quickfix;version="1.0.0";x-friends:="org.eclipse.pde.ui",
  org.eclipse.pde.bnd.ui.templating;version="1.0.0";x-friends:="org.eclipse.pde.ui",
  org.eclipse.pde.bnd.ui.wizards;version="1.0.0";x-friends:="org.eclipse.pde.ui"
 Import-Package: aQute.bnd.build;version="4.5.0",
@@ -40,7 +41,7 @@ Require-Bundle: org.eclipse.jdt.ui,
 Automatic-Module-Name: org.eclipse.pde.bnd.ui
 Bundle-Activator: org.eclipse.pde.bnd.ui.Resources
 Bundle-RequiredExecutionEnvironment: JavaSE-17
+Bundle-ActivationPolicy: lazy
 Service-Component: OSGI-INF/org.bndtools.templating.repos.xml,
  OSGI-INF/org.eclipse.pde.bnd.ui.internal.Auxiliary.xml,
  OSGI-INF/org.eclipse.pde.bnd.ui.internal.TemplateAdapter.xml
-Bundle-ActivationPolicy: lazy
